+++ kernel/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "os"
-version = "0.1.0"
+name = "kernel"
+version = "0.0.1"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
