+++ kernel/src/batch.rs
