+++ kernel/src/console.rs
@@ -12,6 +12,7 @@ impl Write for Stdout {
     }
 }
 
+
 pub fn print(args: fmt::Arguments) {
     Stdout.write_fmt(args).unwrap();
 }
