+++ kernel/src/link_app.S
@@ -15,33 +15,33 @@ _num_app:
     .global app_0_start
     .global app_0_end
 app_0_start:
-    .incbin "../user/target/riscv64gc-unknown-none-elf/release/00hello_world.bin"
+    .incbin "../user_lib/target/riscv64gc-unknown-none-elf/release/00hello_world.bin"
 app_0_end:
 
     .section .data
     .global app_1_start
     .global app_1_end
 app_1_start:
-    .incbin "../user/target/riscv64gc-unknown-none-elf/release/01store_fault.bin"
+    .incbin "../user_lib/target/riscv64gc-unknown-none-elf/release/01store_fault.bin"
 app_1_end:
 
     .section .data
     .global app_2_start
     .global app_2_end
 app_2_start:
-    .incbin "../user/target/riscv64gc-unknown-none-elf/release/02power.bin"
+    .incbin "../user_lib/target/riscv64gc-unknown-none-elf/release/02power.bin"
 app_2_end:
 
     .section .data
     .global app_3_start
     .global app_3_end
 app_3_start:
-    .incbin "../user/target/riscv64gc-unknown-none-elf/release/03priv_inst.bin"
+    .incbin "../user_lib/target/riscv64gc-unknown-none-elf/release/03priv_inst.bin"
 app_3_end:
 
     .section .data
     .global app_4_start
     .global app_4_end
 app_4_start:
-    .incbin "../user/target/riscv64gc-unknown-none-elf/release/04priv_csr.bin"
+    .incbin "../user_lib/target/riscv64gc-unknown-none-elf/release/04priv_csr.bin"
 app_4_end:
