+++ kernel/src/stack_trace.rs
