--- os/src/console.rs
@@ -1,67 +1,0 @@
-use core::fmt::{self, Write};
-
-struct Stdout;
-
-impl Write for Stdout {
-    #[allow(deprecated)]
-    fn write_str(&mut self, s: &str) -> fmt::Result {
-        for c in s.chars() {
-            sbi_rt::legacy::console_putchar(c as usize);
-        }
-        Ok(())
-    }
-}
-
-
-pub fn print(args: fmt::Arguments) {
-    Stdout.write_fmt(args).unwrap();
-}
-
-#[macro_export]
-macro_rules! print {
-    ($fmt: literal $(, $($arg: tt)+)?) => {
-        $crate::console::print(format_args!($fmt $(, $($arg)+)?));
-    }
-}
-
-#[macro_export]
-macro_rules! println {
-    ($fmt: literal $(, $($arg: tt)+)?) => {
-        $crate::console::print(format_args!(concat!($fmt, "\n") $(, $($arg)+)?));
-    }
-}
-
-#[macro_export]
-macro_rules! info {
-    ($fmt: literal $(, $($arg: tt)+)?) => {
-        $crate::console::print(format_args!(concat!("\x1b[32m", "[KERNEL] [INFO   ]: ", $fmt, "\x1b[0m\n") $(, $($arg)+)?));
-    }
-}
-
-#[macro_export]
-macro_rules! debug {
-    ($fmt: literal $(, $($arg: tt)+)?) => {
-        $crate::console::print(format_args!(concat!("\x1b[34m", "[KERNEL] [DEBUG  ]: ", $fmt, "\x1b[0m\n") $(, $($arg)+)?));
-    }
-}
-
-#[macro_export]
-macro_rules! warning {
-    ($fmt: literal $(, $($arg: tt)+)?) => {
-        $crate::console::print(format_args!(concat!("\x1b[93m". "[KERNEL] [WARNING]: ", $fmt, "\x1b[0m\n") $(, $($arg)+)?));
-    }
-}
-
-#[macro_export]
-macro_rules! trace {
-    ($fmt: literal $(, $($arg: tt)+)?) => {
-        $crate::console::print(format_args!(concat!("\x1b[90m", "[KERNEL] [TRACE  ]: ", $fmt, "\x1b[0m\n") $(, $($arg)+)?));
-    }
-}
-
-#[macro_export]
-macro_rules! error {
-    ($fmt: literal $(, $($arg: tt)+)?) => {
-        $crate::console::print(format_args!(concat!("\x1b[31m", "[KERNEL] [ERROR  ]: ", $fmt, "\x1b[0m\n") $(, $($arg)+)?));
-    }
-}
