--- user/.cargo/config
@@ -1,7 +1,0 @@
-[build]
-target = "riscv64gc-unknown-none-elf"
-
-[target.riscv64gc-unknown-none-elf]
-rustflags = [
-    "-Clink-args=-Tsrc/linker.ld",
-]
