--- user/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "user_lib"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-sbi-rt = { version = "0.0.2", features = ["legacy"] }
-riscv = { git = "https://github.com/rcore-os/riscv", features = ["inline-asm"] }
-
-[profile.release]
-debug = true
