--- user/src/bin/00hello_world.rs
@@ -1,11 +1,0 @@
-#![no_std]
-#![no_main]
-
-#[macro_use]
-extern crate user_lib;
-
-#[no_mangle]
-fn main() -> i32 {
-    println!("Hello, world!");
-    0
-}
