--- user/src/bin/04priv_csr.rs
@@ -1,17 +1,0 @@
-#![no_std]
-#![no_main]
-
-#[macro_use]
-extern crate user_lib;
-
-use riscv::register::sstatus::{self, SPP};
-
-#[no_mangle]
-fn main() -> i32 {
-    println!("Try to access privileged CSR in U Mode");
-    println!("Kernel should kill this application!");
-    unsafe {
-        sstatus::set_spp(SPP::User);
-    }
-    0
-}
