--- user/src/linker.ld
@@ -1,30 +1,0 @@
-OUTPUT_ARCH(riscv)
-ENTRY(_start)
-BASE_ADDRESS = 0x80400000;
-
-SECTIONS
-{
-    . = BASE_ADDRESS;
-    .text : {
-        *(.text.entry)
-        *(.text .text.*)
-    }
-    .rodata : {
-        *(.rodata .rodata.*)
-        *(.srodata .srodata.*)
-    }
-    .data : {
-        *(.data .data.*)
-        *(.sdata .sdata.*)
-    }
-    .bss : {
-        start_bss = .;
-        *(.bss .bss.*)
-        *(.sbss .sbss.*)
-        end_bss = .;
-    }
-    /DISCARD/ : {
-        *(.eh_frame)
-        *(.debug*)
-    }
-}
